perf(client): split translation state from on-demand helpers
